requirements.txt
fastapi>=0.135.0
uvicorn[standard]
asyncpg>=0.29.0
httpx>=0.27.0
pydantic>=2.12.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
openai>=2.0.0
twilio>=9.0.0
python-dateutil>=2.8.0
pytz
python-multipart
faster-whisper>=1.0.0
